using delete action

+++ lightningwolf_smp/blueprints/user.py
@@ -31,13 +31,19 @@ def user_page():
 @user.route('/admin/user/list', methods=["GET"])
 @admin_permission.require(http_exception=403)
 def user_list():
+    from flask_wtf import Form
     from lightningwolf_smp.utils.user import get_user_list
     from lightningwolf_smp.forms.user import FormUsernameFilter, FormUserBatchActions
-    filter = FormUsernameFilter()
-    batch_actions = FormUserBatchActions()
     users = get_user_list()
     navbar = create_navbar_fd(navbar_conf, 'key.user.user_list')
-    return render_template('user/list.html', lw_navbar=navbar, list=users, filter=filter, batch_actions=batch_actions)
+    return render_template(
+        'user/list.html',
+        lw_navbar=navbar,
+        list=users,
+        filter=FormUsernameFilter(),
+        batch_actions=FormUserBatchActions(),
+        delete_action=Form()
+    )
 
 
 @user.route('/admin/user/filter', methods=["POST"])
@@ -114,9 +120,15 @@ def user_edit(id):
 @user.route('/admin/user/<int:id>/delete', methods=["POST"])
 @admin_permission.require(http_exception=403)
 def user_del(id):
+    from flask_wtf import Form
     from lightningwolf_smp.utils.user import get_user
     user = get_user(id)
     if user is None:
         abort(404)
-    user.delete()
+    form = Form()
+    if form.validate_on_submit():
+        user.delete()
+    else:
+        flash(u'An error occurred while deleting the user', 'error')
+
     return redirect(url_for('user.user_list'))
